#3 udpated duration rendering and data visualisation

+++ src/main/java/at/mana/idea/domain/MethodEnergyStatisticsSample.java
@@ -12,7 +12,7 @@ import java.util.Arrays;
 @AllArgsConstructor
 public class MethodEnergyStatisticsSample {
     
-    private long duration;
+    private double duration;
     private Double[] cpu;
     private Double[] gpu;
     private Double[] ram;
